👷(i18n) implement a hack to serve fr pages on root
As we're deploying 100% static HTML, we cannot do server based
redirects based on headers. We need the language for 95% of our
users to live at root, and alternative languages on their respective
/{lang}/ path.

`eleventy-plugin-i18n`, which was used to build funlandpage, does
not support building/serving one default language on the root "/"
path.

We managed to achieve it by making the defaut language (fr) home
be served at the root of the project, and setting the default
language as the fallback for localization.

We also had to change the localized site data to regular strings
so they get the benefit of the i18n filter and therefore the
fallback locale.