🐛(portfolio) manage portfolio items active states
When the user clicks on items in the portfolio, it swipes to the
item in question and adds an active state on its button. When the
user just swipes through the items, the buttons' active states
were not updated.

Add a hook on swipes so button active states are updated when the
user swipes through.