✨(i18n) redirect non french speakers to the english home
As we're serving this website 100% statically, we cannot use headers
to return the correct language from the beginning. We decided to use
the french page as home as we suspect most users will we looking for
the french version.

To make things easier for non french speakers, we decided to add a
frontend JS redirect to the english page if french is not detected
among the navigator languages. We then duplicated the french version
at "/fr/" (in addition to "/") to have a place where we send users who
deliberately picked the french version and not redirect them.